requirements.txt
streamlit==1.52.2
streamlit-option-menu

requests
tavily-python
pandas
altair
python-dotenv

langchain
langchain-groq
langchain-community
langchain-text-splitters
langchain-huggingface

faiss-cpu
sentence-transformers
huggingface-hub

pypdf
python-docx
openpyxl

bcrypt==5.0.0

google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3

Authlib==1.6.6
